+++ bazel/ios/rewrite_symbols.sh
@@ -2,9 +2,9 @@
 
 set -euxo pipefail
 
-AR=$(which llvm-ar)
-RANLIB=$(which llvm-ranlib)
-LIPO=$(which llvm-lipo)
+: "${AR:?AR is not set or empty}"
+: "${RANLIB:?RANLIB is not set or empty}"
+: "${LIPO:?LIPO is not set or empty}"
 
 # Ensure that ar/ranlib creates deterministic archives.
 export ZERO_AR_DATE=1