+++ platform/jvm/capture/src/test/kotlin/io/bitdrift/capture/events/lifecycle/AppExitLoggerTest.kt
@@ -62,6 +62,7 @@ class AppExitLoggerTest {
     @Before
     fun setUp() {
         whenever(runtime.isEnabled(RuntimeFeature.APP_EXIT_EVENTS)).thenReturn(true)
+        whenever(runtime.isEnabled(RuntimeFeature.LOGGER_FLUSHING_ON_CRASH)).thenReturn(true)
         whenever(versionChecker.isAtLeast(anyInt())).thenReturn(true)
         appExitLogger = buildAppExitLogger()
         lastExitInfo.reset()
@@ -183,7 +184,6 @@ class AppExitLoggerTest {
     @Test
     fun testHandlerCrashLogs() {
         // ARRANGE
-        whenever(runtime.isEnabled(RuntimeFeature.LOGGER_FLUSHING_ON_CRASH)).thenReturn(true)
         val currentThread = Thread.currentThread()
         val appException = IOException("real app crash")
 
@@ -212,7 +212,6 @@ class AppExitLoggerTest {
             eq(true),
             argThat { i: () -> String -> i.invoke() == "AppExit" },
         )
-        verify(logger).flush(true)
     }
 
     @Test
@@ -237,7 +236,6 @@ class AppExitLoggerTest {
     @Test
     fun onJvmCrash_whenBuiltInFatalIssueMechanism_shouldNotSendAppExitCrashLog() {
         val appExitLogger = buildAppExitLogger(IssueReporterState.Initialized)
-        whenever(runtime.isEnabled(RuntimeFeature.LOGGER_FLUSHING_ON_CRASH)).thenReturn(true)
 
         appExitLogger.onJvmCrash(Thread.currentThread(), IllegalStateException("Simulated Crash"))