+++ platform/jvm/gradle-test-app/src/main/java/io/bitdrift/gradletestapp/data/model/Actions.kt
@@ -102,6 +102,8 @@ sealed class NavigationAction : AppAction {
 
     object NavigateToStressTest : NavigationAction()
 
+    object NavigateToFocusMatrix : NavigationAction()
+
     object InvokeService : NavigationAction()
 }