+++ platform/jvm/gradle-test-app/src/main/java/io/bitdrift/gradletestapp/ui/viewmodel/MainViewModel.kt
@@ -197,6 +197,7 @@ class MainViewModel(
             is NavigationAction.NavigateToXml -> {}
             is NavigationAction.NavigateToDialogAndModals -> {}
             is NavigationAction.NavigateToStressTest -> {}
+            is NavigationAction.NavigateToFocusMatrix -> {}
             is NavigationAction.InvokeService -> {}
 
         }