+++ test/benchmark/src/bin/live_benchmark.rs
@@ -5,15 +5,15 @@
 // LICENSE file or at:
 // https://polyformproject.org/wp-content/uploads/2020/06/PolyForm-Shield-1.0.0.txt
 
-use bd_logger::{log_level, AnnotatedLogField, Block, CaptureSession, InitParams};
+use bd_logger::{AnnotatedLogField, CaptureSession, InitParams, log_level};
 use bd_proto::protos::logging::payload::LogType;
-use bd_session::fixed::UUIDCallbacks;
 use bd_session::Strategy;
+use bd_session::fixed::UUIDCallbacks;
 use bd_shutdown::ComponentShutdownTrigger;
 use bd_test_helpers::metadata::EmptyMetadata;
 use bd_test_helpers::metadata_provider::LogMetadata;
 use bd_test_helpers::session::InMemoryStorage;
-use criterion::{criterion_group, criterion_main, Criterion};
+use criterion::{Criterion, criterion_group, criterion_main};
 use std::sync::Arc;
 use std::time::Duration;
 
@@ -40,9 +40,10 @@ fn test_live_match_performance(c: &mut Criterion) {
     network: Box::new(network),
     static_metadata: Arc::new(EmptyMetadata),
     sdk_directory: ".".into(),
-    session_strategy: Arc::new(Strategy::fixed(".", Arc::new(UUIDCallbacks))),
+    session: Strategy::fixed(".", Arc::new(UUIDCallbacks)),
     store,
     metadata_provider,
+    initial_ootb_fields: [].into(),
     resource_utilization_target: Box::new(bd_test_helpers::resource_utilization::EmptyTarget),
     session_replay_target: Box::new(bd_test_helpers::session_replay::NoOpTarget),
     events_listener_target: Box::new(bd_test_helpers::events::NoOpListenerTarget),
@@ -66,7 +67,6 @@ fn test_live_match_performance(c: &mut Criterion) {
         [].into(),
         [].into(),
         None,
-        Block::No,
         &CaptureSession::default(),
       );
       handle.log(
@@ -76,7 +76,6 @@ fn test_live_match_performance(c: &mut Criterion) {
         [].into(),
         [].into(),
         None,
-        Block::No,
         &CaptureSession::default(),
       );
       handle.log(
@@ -86,7 +85,6 @@ fn test_live_match_performance(c: &mut Criterion) {
         [].into(),
         [].into(),
         None,
-        Block::No,
         &CaptureSession::default(),
       );
       handle.log(
@@ -96,7 +94,6 @@ fn test_live_match_performance(c: &mut Criterion) {
         [].into(),
         [].into(),
         None,
-        Block::No,
         &CaptureSession::default(),
       );
       handle.log(
@@ -106,7 +103,6 @@ fn test_live_match_performance(c: &mut Criterion) {
         [].into(),
         [].into(),
         None,
-        Block::No,
         &CaptureSession::default(),
       );
     });
@@ -129,7 +125,6 @@ fn test_live_match_performance(c: &mut Criterion) {
         .into(),
         [].into(),
         None,
-        Block::No,
         &CaptureSession::default(),
       );
     });